Rework model defaults get, set & unset per new spec
Refs: juju-model-config-collapse
Refs: juju-model-defaults-collapse
