+++ cmd/juju/commands/main_test.go
@@ -491,7 +491,6 @@ var commandNames = []string{
 	"set-default-region",
 	"set-meter-status",
 	"set-model-constraints",
-	"set-model-default",
 	"set-plan",
 	"shares",
 	"show-action-output",
@@ -519,7 +518,6 @@ var commandNames = []string{
 	"update-allocation",
 	"upload-backup",
 	"unregister",
-	"unset-model-default",
 	"update-clouds",
 	"upgrade-charm",
 	"upgrade-gui",
